+++ agents/relayer/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- relayer checks replica updater addresses match, errors channel if otherwise
 - add bootup-only tracing subscriber
 - add environment variable overrides for agent configuration
 - add tests for agent environment variable overrides